requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
